memory fs: add set/get current directory; refactored code
Added support of current directory. With this feature the behaviour of
makedirs() is somewhat changed, thus it's updated as needed. Introduced
path normalization, and further checks, which has more sense if there
is a current directory. The test class tests only these.

Currently FileExistsError and FileNotFoundError exceptions are raised,
exactly as in case of the native methods, os.makedirs and os.chdir.